% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_representations_classes.R
\name{as.stspmod}
\alias{as.stspmod}
\alias{as.stspmod.armamod}
\title{Coerce to State Space Model}
\usage{
as.stspmod(obj, ...)

\method{as.stspmod}{armamod}(obj, ...)
}
\arguments{
\item{obj}{object}

\item{...}{optional additional parameters}

\item{method}{character string}
}
\value{
object of class \code{\link[=stspmod]{stspmod()}}.
}
\description{
The function \code{\link[=as.stsp.pseries]{as.stsp.pseries()}} calls
\code{\link[=pseries2stsp]{pseries2stsp()}} with default parameters.
Of course the \code{\link[=pseries]{pseries()}} object must contain
sufficiently many lags. NOT YET implemented
}
\examples{
# Convert an ARMA model to state space representation
arma_model = armamod(sys = lmfd(c(1, 0.5, 0.2), c(1, -0.3)), sigma_L = diag(1))
ss_model = as.stspmod(arma_model)
ss_model

}
