% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BQdata}
\alias{BQdata}
\alias{BQdata_ts}
\alias{BQdata_xts}
\title{Blanchard/Quah (1989) dataset}
\format{
A tibble (BQdata), a ts-object (BQdata_ts), or an xts-object (BQdata_xts) with 159 rows and 2 variables (plus a timestamp, so in total 3):
\describe{
\item{date}{Timestamp as column of type \emph{dttm} for the tibble, and as index for the \emph{ts} and \emph{xts} object.}
\item{rGDPgrowth_demeaned}{The real GDP growth series is demeaned with respect to two different subperiods: Till the last quarter of 1973 and from the first quarter of 1974}
\item{unemp_detrended}{The unemployment rate is detrended with respect to a linear trend (and an intercept).}
}

An object of class \code{mts} (inherits from \code{ts}, \code{matrix}) with 159 rows and 2 columns.

An object of class \code{xts} (inherits from \code{zoo}) with 159 rows and 2 columns.
}
\source{
\url{https://academic.oup.com/restud/advance-article/doi/10.1093/restud/rdz028/5490841#supplementary-data}
}
\usage{
BQdata

BQdata_ts

BQdata_xts
}
\description{
This dataset contains a (159 x 2)-dimensional matrix with quarterly data of real GDP growth rates (first column) and the unemployment rate in the USA.
It starts in the second quarter of 1948 and ends in the last quarter of 1987.
The script for transforming the raw data (as csv-file) to the matrix BQdata is available in the \emph{data-raw} directory.
This dataset was used in \href{https://doi.org/10.1093/restud/rdz028}{Gourieroux, Monfort, Renne (2019)}
}
\keyword{datasets}
