% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cll_theta_STSP_cpp}
\alias{cll_theta_STSP_cpp}
\title{Compute the (concentrated) conditional log likelihood for a statespace system
described by a model template.}
\arguments{
\item{th}{\eqn{(K)} dimensional vector of "deep" parameters.}

\item{y}{\eqn{(m,N)} matrix with the observed outputs:
\eqn{(y_1,y_2,\ldots,y_N)}{(y[1],y[2],...,y[N])}.}

\item{skip}{(integer), skip the first residuals, when computing the sample covariance of the
residuals.}

\item{concentrated}{(bool), if TRUE then the \emph{concentrated}, conditional log Likelihood is computed}

\item{pi}{\eqn{(m+s,m+s)} matrix, is overwritten with the system matrix
\eqn{[A,B | C,D]}.}

\item{H_pi}{\eqn{(m+s)^2, K)} matrix.}

\item{h_pi}{\eqn{((m+s)^2)}-dimensional vector. Note that \code{vec(pi) = H_pi*th + h_pi}.}

\item{L}{\eqn{(m,m)} matrix. If (concentrated==FALSE) then L is overwritten with
the left square of the noise covariance matrix L corresponding
to the deep parameters th. However, if (concentrated==TRUE) then
L is overwritten with sample covariance matrix of the computed residuals!}

\item{H_L}{\eqn{(m^2, K)} matrix.}

\item{h_L}{\eqn{(m^2)}-dimensional vector. Note that
\code{vec(L) = H_L*th + h_L}.}

\item{a}{\eqn{(s,N+1)} matrix. This matrix is overwritten with the (computed) states:
\eqn{(a_1,a_2,\ldots,a_N,a_{N+1})}{(a[1],a[2],\ldots,a[N],a[N+1])}.
On input \code{a[,1]} must hold the initial state \eqn{a_1}{a[1]}.}

\item{u}{\eqn{(m,N)} matrix. This matrix is overwritten with (computed) residuals:
\eqn{(u_1,u_2,\ldots,u_N)}{(u[1],u[2],...,u[N])}.}

\item{dU}{\eqn{(mN,K)} matrix or \eqn{(0,0)} matrix. This matrix is overwritten with the
directional derivatives of the residuals. However, if
the matrix is empty then no derivatives are computed.}
}
\value{
(double) log Likelihood
}
\description{
This is an internal helper function, used by the function factory \code{\link{ll_FUN}}. For a more detailed
documentation of the conditional log Likelihood, see \code{\link{ll}}.
The conditional likelihood is computed for the initial state \eqn{a_1} given in the first column \code{a[,1]} of the
matrix \code{a}.
}
\seealso{
\code{\link{outputs_ARMA_cpp}}, \code{\link{residuals_ARMA_cpp}},  \code{\link{cll_theta_ARMA_cpp}},
\code{\link{outputs_STSP_cpp}}, \code{\link{residuals_STSP_cpp}},  \code{\link{cll_theta_STSP_cpp}} and
\code{\link{solve_de}}, \code{\link{solve_inverse_de}} and \code{\link{ll}}.
}
