% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afab_estimate_mom_arma.R
\name{est_arma_hrk3}
\alias{est_arma_hrk3}
\title{Different version of HRK Procedure}
\usage{
est_arma_hrk3(
  y,
  tmpl,
  maxit_stage2 = 5,
  tol_stage2 = 0.001,
  maxit_stage3 = 5,
  tol_stage3 = 0.001,
  info = TRUE,
  trace = FALSE,
  p.max = NULL,
  ic = c("AIC", "BIC", "max"),
  mean_estimate = c("zero", "sample.mean", "intercept"),
  tol = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{y}{sample, i.e. an \eqn{(N,m)} dimensional matrix,
or a "time series" object (i.e. \code{as.matrix(y)} should return an
\eqn{(N,m)}-dimensional numeric matrix). Missing values (\code{NA}, \code{NaN} and
\code{Inf}) are \strong{not} supported.}

\item{tmpl}{a model template, see \code{\link[=model structures]{model structures()}}. Note that only the case
is implemented, where
\eqn{a_0=b_0}{a[0]=b[0]} holds, the diagonal entries of
\eqn{a_0=b_0}{a[0]=b[0]} are equal to one and all other fixed elements
are equal to zero. Furthermore the square root \code{sigma_L} of the noise covariance matrix
is asssumed to be a lower triangular matrix without any further restrictions.
\cr
The given template is coerced to a template of this kind. If the given template does not
comply to these restrictions, then a warning message is issued.}

\item{maxit_stage2, maxit_stage3}{Integers. Default for both stages is 5.}

\item{tol_stage2, tol_stage3}{Default set to \code{1e-3}.
Maximal absolute distance of the deep parameters of adjacent iterations}

\item{info}{Boolean. Indicates whether the slot \code{extra_info} of the returned list
should contain a tibble with additional info about the model at different iterations,
e.g., min and max absolute value of zeros, poles, value of objective function etc.}

\item{trace}{(boolean) if \code{trace=TRUE}, then some tracing information on the
iterations is printed.}

\item{p.max}{(integer or \code{NULL}) Maximum order of the candidate AR models.
For the default choice see below.}

\item{ic}{(character string) Which information criterion shall be used to find the
optimal AR order. Note that \code{ic="max"} means that an AR(p) model
with \code{p=p.max} is fitted. Default is \code{ic="AIC"}.}

\item{mean_estimate}{Character string giving the method used to estimate the mean \eqn{\mu}.
Default is \code{mean_estimate = "sample.mean"}.
See the details below.}

\item{tol}{Small tolerance, used to check whether the mean of the supplied data matrix is indeed zero.}
}
\value{
See \code{\link[=est_arma_hrk]{est_arma_hrk()}}.
The list contains the additional slots
\describe{
\item{stage_opt}{Since the best stable and miniphase model is returned, we also indicate whether this happened at stage II or stage III.}
\item{info_tibble}{Tibble containing all relevant info about the outcome at different stages and iterations.}
}
and does not contain the slots \code{y.mean} (since mean is required to be zero) and \code{converged}.
}
\description{
See \link{est_arma_hrk}.
Stage III of the Hannan-Rissanen-Kavalieris procedure is implemented here as well.
This function returns the best model (since the iterations might not always improve the log-likelihood value) and
allows for returning results at different stages of the HRK procedure.
\cr
One notable differences is that the data needs to be demeaned because
we use Yule-Walker estimation in the first stage to ensure stability and because
the implementation of stage III would otherwise be even more cumbersome.
}
\examples{
data = BQdata_xts
for (pp in 0:2){
  for (qq in 0:2){
    if (pp + qq == 0){next}
    cat(paste0("p = ", pp, " q = ", qq, "\n"))
    tmpl = tmpl_arma_pq(m = 2, n = 2,
                        p = pp, q = qq)
    ff = est_arma_hrk3(data, tmpl = tmpl)
  }
}
ff$info_tibble \%>\% print(n=100)
}
