% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fbsolve_STSP_cpp}
\alias{fbsolve_STSP_cpp}
\title{Forward-backward solution of statespace systems}
\arguments{
\item{A}{\eqn{(s, s)} matrix.}

\item{B}{\eqn{(s, n)} matrix.}

\item{C}{\eqn{(m, s)} matrix.}

\item{D}{\eqn{(m, n)} matrix.}

\item{u}{\eqn{(n, N)} matrix with the inputs \eqn{(u_1,...,u_N}{(u[1],...,u[N])}.}

\item{au}{\eqn{(su,N+1)} matrix. This matrix is \strong{overwritten} with the (computed)
states of the unstable part of the system.
\eqn{(a_{u1},a_{u2},\ldots,a_{uN},a_{u,N+1})}{(au[1],au[2],\ldots,au[N],au[N+1])}.
On input \code{au[,N+1]} must hold the "initial" state \eqn{a_{u,N+1}}{au[N+1]}.}

\item{as}{\eqn{(ss,N+1)} matrix. This matrix is \strong{overwritten} with the (computed)
states of the stable part of the system.
\eqn{(a_{s1},a_{s2},\ldots,a_{sN},a_{s,N+1})}{(as[1],as[2],\ldots,as[N],as[N+1])}.
On input \code{as[,1]} must hold the "initial" state \eqn{a_{s1}}{as[1]}.}

\item{y}{\eqn{(m,N)} matrix. This matrix is \strong{overwritten} with (computed) outputs:
\eqn{(y_1,y_2,\ldots,y_N)}{(y[1],y[2],...,y[N])}.}
}
\value{
This RcppArmadillo routine returns \code{NULL} but \bold{overwrites}
the input argument \code{y}, \code{au} and \code{as} with the computed outputs
and states!
}
\description{
\strong{DEPRECATED}?
This internal helper function computes the outputs of an, in general \strong{unstable}, statespace system
\deqn{a_{t+1} = A a_t + B u_t, \; y_t = C a_t + D u_t}{
      a[t+1] = A a[t] + B u[t] and y[t] = C a[t] + D u[t]}
by forward and backward recursion. The procedure assumes that the state transition matrix
\eqn{A} is block upper triangular, where the upper block \eqn{A_{11}}{A[11]} is stable (i.e.
all eigenvalues have moduli less than one) and the
lower block \eqn{A_{22}}{A[22]} is unstable (i.e. all eigenvalues have moduli larger than one).
This function is mainly used in the routine \code{\link{innovation_form}}.
}
\note{
Use this procedure with care!
\itemize{
\item The procedure does \bold{not} check the input arguments. We require \eqn{m > 0}, \eqn{n > 0}.
Furthermore it is assumed that the state transition matrix \eqn{A} is block upper triangular,
as explained above.
\item The procedure \bold{overwrites} the input arguments \code{y}, \code{as} and \code{au}.
\item The data matrices are organized columnwise (to avoid memory shuffling)!
}
}
\seealso{
\code{\link{outputs_STSP_cpp}} and \code{\link{innovation_form}}.
}
