% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc_impresp_methods.R
\name{fevardec}
\alias{fevardec}
\title{Forecast Error Variance Decomposition}
\usage{
fevardec(obj, h.max = NULL, H = NULL)
}
\arguments{
\item{obj}{\code{\link[=impresp]{impresp()}} object which represents the (orthogonalized) impulse response function.}

\item{h.max}{maximum forecast horizon. The default is one plus the number of lags of the \code{impresp}
object.}

\item{H}{An (n x n) (non singular) matrix which renders the impulse response to a
\emph{orthogonalized} impulse response. The noise \eqn{u_t}{u[n]} is transformed to
\eqn{H^{-1}u_t}{H^{-1}u[t]} and the impulse response coefficients
(\eqn{k_j \rightarrow k_j H}{k[j] -> k[j] H}) and the (left)
square root of the noise covariance matrix (\eqn{L \rightarrow H^{-1}L}{L -> H^{-1}L}) are
transformed correspondingly.
\cr
The default case \code{H=NULL} corresponds to the identity matrix (i.e. no transformation).
\cr
For \code{H='chol'}, the transformation matrix \code{H = t(chol(Sigma))}
is obtained from the Choleski decomposition of the noise covariance \eqn{\Sigma}.
For \code{H='eigen'} the symmetric square root of
\eqn{\Sigma} (obtained from the the eigenvalue decomposition of \eqn{\Sigma})
is used. For \code{H='sigma_L'} the left square root of the noise covariance,
which is stored in the object \code{obj}, is used.
Other orthogonalization schemes may be obtained by setting \eqn{H} to a
suitable square root of \eqn{\Sigma}.
\cr
The procedure checks whether the transformation yields an orthogonalized
impulse response. If not, an error is thrown.}
}
\value{
\code{fevardec} object, i.e. a list with components
\item{vd}{n-by-n-by-h.max array which contains the forecast error variance
decomposition: \code{vd[i,j,h]} is the percentage of the variance of the
h-step ahead forecast error of the i-th component due to the j-th
orthogonalized shock.}
\item{v}{n-by-h.max matrix which contains the forecast error variances:
\code{v[i,h]} is the variance of the h-step ahead forecast error for the
i-th component.}
\item{names}{(m)-dimensional character vector}
\item{label}{character string or NULL}
}
\description{
Computes the Forecast Errors Variance Decomposition from a given (orthogonalized)
impulse response function.
}
\examples{
model = stspmod(sys = stsp(A = c(0,0.2,1,-0.5), B = c(1,1,1,-1),
                           C = c(1,0,0,1)),
                sigma_L = t(chol(matrix(c(4,2,2,3),nrow=2))),
                names = c('y1','y2'), label = 'test model')
fevardec(impresp(model, lag.max=10), H = 'chol', h.max = 5) \%>\% print(digits = 2, format = 'iz|j')
fevardec(impresp(model, lag.max=4, H = 'eigen'))            \%>\% print(digits = 2, format = 'iz|j')
}
