% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_templates.R
\name{is.template}
\alias{is.template}
\title{Check templates}
\usage{
is.template(tmpl, class = c("any", "stspmod", "armamod", "rmfdmod"))
}
\arguments{
\item{tmpl}{object to be tested}

\item{class}{test for a specific class}
}
\value{
TRUE/FALSE
}
\description{
Check templates
}
\examples{
is.template(1)
is.template(tmpl_llm(), 'armamod')
is.template(tmpl_llm(), 'stspmod')
}
