% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_compare_models.R
\name{KL_divergence}
\alias{KL_divergence}
\title{Kullback–Leibler divergence}
\usage{
KL_divergence(model, model_hat)
}
\arguments{
\item{model}{\code{\link[=stspmod]{stspmod()}} object, true model}

\item{model_hat}{\code{\link[=stspmod]{stspmod()}} object, estimated model}
}
\value{
KL divergence.
}
\description{
Compute the Kullback-Leibler divergence between a "true" state space model and
an estimated state space model. The function only works for square systems,
where the "true" model is stable and the estimate is (strictly) miniphase.
}
\details{
The KL divergence is computed as follows. Suppose \eqn{y_t = k(z) u_t}{y[t] = k(z) u[t]},
with \eqn{\mathbf{E} u_t u_t' = \Sigma}{E u[t] u[t]' = \Sigma} is the true model, and
let \eqn{y_t = h(z) u_t}{y[t] = h(z) u[t]}, with \eqn{\mathbf{E} u_t u_t'=\Omega}{E u[t] u[t]'=\Omega}
denote the estimate. W.l.o.g. we assume that the models are in innovation form,
i.e. \eqn{k(0) = h(0) = I}. The procedure computes the covariance matrix,
\eqn{\Delta = \mathbf{E} e_t e_t'}{\Delta = E e[t] e[t]'} say,
of the \emph{one-step-ahead prediction errors} \eqn{e_t = h^{-1}(z) k(z) u_t}{e[t] = h^{-1} (z) k(z) u[t]}
and then the KL divergence
\deqn{
\mathrm{KL} = (1/2)(\mathrm{tr}(\Omega^{-1}\Delta) - m - \ln\det(\Omega^{-1}\Delta)))
}{
KL = (1/2)(tr(\Omega^{-1}\Delta) - m - ln det(\Omega^{-1}\Delta)))
}
Note that this procedure breaks down if the transfer function \eqn{h^{-1}(z) k(z)} is not stable.
Therefore the true models has to be stable and the estimated model has to strictly miniphase.
}
