% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ll_kf_theta_cpp}
\alias{ll_kf_theta_cpp}
\title{Compute the log likelihood for a statespace system
described by a model template.}
\usage{
ll_kf_theta_cpp(
  theta,
  y,
  SYS,
  H_SYS,
  h_SYS,
  sigma_L,
  H_sigma_L,
  h_sigma_L,
  VAR,
  P1,
  tol,
  err
)
}
\arguments{
\item{theta}{\eqn{(K)} dimensional vector of "deep" parameters.}

\item{y}{\eqn{(m,N)} matrix with the observed outputs:
\eqn{(y_1,y_2,\ldots,y_N)}{(y[1],y[2],...,y[N])}.}

\item{SYS}{\eqn{(m+s,m+s)} matrix, is overwritten with the system matrix
\eqn{[A,B | C,D]}.}

\item{H_SYS}{\eqn{(m+s)^2, K)} matrix.}

\item{h_SYS}{\eqn{((m+s)^2)}-dimensional vector. Note that \code{vec(SYS) = H_SYS*theta + h_SYS}.}

\item{sigma_L}{\eqn{(m,m)} matrix, is overwritten with the left square root of the
noise covariance matrix.}

\item{H_sigma_L}{\eqn{(m^2, K)} matrix.}

\item{h_sigma_L}{\eqn{(m^2)}-dimensional vector. Note that
\code{vec(sigma_L) = H_sigma_L*theta + h_sigma_L}.}

\item{VAR}{\eqn{(m+s,m+s)} matrix, is overwritten with the
covariance matrix \eqn{[Q,S | S',R] = [B | C] sigma_L sigma_L' [B', C']}}

\item{P1}{\eqn{(s,s)} matrix, is overwritten with the
initial state covariance matrix (computed via a Lyapunov equation).}

\item{tol}{(double) tolerance used by ll_kf_cpp.}

\item{err}{(double) return err, if the computation of P1 fails.}
}
\description{
This is an internal helper function, used by the function factory \code{\link{ll_FUN}}. For a more detailed
documentation of the log Likelihood, see \code{\link{ll_kf}}.
}
