% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_estimation_likelihood.R
\name{ll_theta}
\alias{ll_theta}
\title{Log-likelihood Given Deep Parameters}
\usage{
ll_theta(th, template, y, which, ...)
}
\arguments{
\item{th}{Vector of deep parameter}

\item{template}{A model template, see \link{model structures}.}

\item{y}{Data sample given as \eqn{(N,m)} dimensional matrix, or
a "time series" object (in the sense that \code{as.matrix(y)}
should return an \eqn{(N,m)}-dimensional numeric matrix).
Missing values (\code{NA}, \code{NaN} and \code{Inf}) are \strong{not} supported.}

\item{which}{(character) which likelihood to compute.}

\item{...}{Not used.}
}
\value{
Value of log-likelihood for a given deep/free parameter vector \code{th} and
a model structure defined via \code{template}.
Note that this function simply calls \code{ll(fill_template(th, template), y, which, ...)}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
See \code{\link[=ll_FUN]{ll_FUN()}}.
The template \code{template} is filled with the deep parameters in \code{th}.
Subsequently, the S3 method \code{\link[=ll]{ll()}} is called for the class provided in the template
and the value of the \strong{scaled} log-likelihood function is returned, see \code{\link[=ll]{ll()}}.
}
