% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_compare_models.R
\name{pm_test}
\alias{pm_test}
\title{Portmanteau Test for Serial Correlation}
\usage{
pm_test(u, lag.max, n.par)
}
\arguments{
\item{u}{(N-by-m) matrix of residuals (or an object which may be coerced to
a matrix with \code{as.matrix(u)}).}

\item{lag.max}{(integer) maximum number of lags.}

\item{n.par}{(integer) number of parameters of the estimated model.}
}
\value{
Matrix with four columns ("\code{lags}" number of lags, "\code{df}" degrees of freedom,
"\code{Q}" test statistics  and "\code{p}" p values).
}
\description{
Test whether the residuals of an estimated model are serially correlated. The test statistic is
\deqn{Q = N^2\sum_{k=1}^{K} (N-k)^{-1}\mbox{tr} (G_k G_0^{-1} G_k' G_0^{-1})}{
      Q = N^2\sum_{k=1}^{K} (N-k)^{-1} tr (G[k] G[0]^{-1} G[k]' G[0]^{-1})}
where \eqn{G_k}{G[k]} are the sample covariances of the residuals. Under the Null of a correctly
specified and estimated model the test statistic is asmyptotically Chi-squared distributed with
\eqn{Km^2-\kappa} degrees of freedom, where \eqn{\kappa} is the number of (free) parameters of the
model (class).
}
\examples{
u = matrix(rnorm(100*3), nrow = 100, ncol = 3)
pm_test(u, 4, 0)
}
