% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ada_poles_and_zeroes.R
\name{poles and zeroes}
\alias{poles and zeroes}
\alias{zeroes.armamod}
\alias{poles.armamod}
\alias{zeroes.rmfdmod}
\alias{poles.rmfdmod}
\alias{zeroes.stspmod}
\alias{poles.stspmod}
\title{Poles and Zeroes}
\usage{
\method{zeroes}{armamod}(x, tol = sqrt(.Machine$double.eps), print_message = TRUE, ...)

\method{poles}{armamod}(x, tol = sqrt(.Machine$double.eps), print_message = TRUE, ...)

\method{zeroes}{rmfdmod}(x, tol = sqrt(.Machine$double.eps), print_message = TRUE, ...)

\method{poles}{rmfdmod}(x, tol = sqrt(.Machine$double.eps), print_message = TRUE, ...)

\method{zeroes}{stspmod}(x, tol = sqrt(.Machine$double.eps), print_message = TRUE, ...)

\method{poles}{stspmod}(x, tol = sqrt(.Machine$double.eps), print_message = TRUE, ...)
}
\arguments{
\item{x}{an object which represents a VARMA, RMFD or statespace model
(i.e. an \code{\link[=armamod]{armamod()}}, \code{\link[=rmfdmod]{rmfdmod()}} or \code{\link[=stspmod]{stspmod()}} object).}

\item{tol}{Double. Default set to \code{sqrt(.Machine$double.eps)}.
Required to decide on when a root is to be considered "at infinity".}

\item{print_message}{Boolean. Default set to TRUE.
Prints a message if roots "at infinity " are discarded.}

\item{...}{not used.}
}
\value{
Vector of poles, respectively zeroes.
}
\description{
Compute the poles and zeroes of VARMA and Statespace models. Note that these models describe
the corresponding processes as
\deqn{x_t = k(B) u_t}{x[t] = k(B) + u[t]}
where \eqn{(u_t)}{(u[t])} is a white noise process and \eqn{k(B)} is a rational filter
(\eqn{B} denotes the lag- or backward shift operator).
The poles and zeroes are the poles and zeroes of the rational transfer function
\eqn{k(z)} of this filter.
}
\seealso{
For more details we refer to the discussion about the computation of poles and zeros
of rational matrices in the companion package,
see \link[rationalmatrices:poles_and_zeroes]{rationalmatrices::poles and zeroes}.
}
