% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_utilities_print.R
\name{print methods}
\alias{print methods}
\alias{print.armamod}
\alias{print.rmfdmod}
\alias{print.stspmod}
\alias{print.impresp}
\alias{print.autocov}
\alias{print.fevardec}
\alias{print.freqresp}
\alias{print.spectrald}
\title{Print Methods}
\usage{
\method{print}{armamod}(
  x,
  digits = NULL,
  format = c("i|jz", "i|zj", "iz|j", "zi|j", "i|j|z", "character"),
  ...
)

\method{print}{rmfdmod}(
  x,
  digits = NULL,
  format = c("i|jz", "i|zj", "iz|j", "zi|j", "i|j|z", "character"),
  ...
)

\method{print}{stspmod}(x, digits = NULL, ...)

\method{print}{impresp}(
  x,
  digits = NULL,
  format = c("i|jz", "i|zj", "iz|j", "zi|j", "i|j|z"),
  ...
)

\method{print}{autocov}(
  x,
  digits = NULL,
  format = c("i|jz", "i|zj", "iz|j", "zi|j", "i|j|z"),
  ...
)

\method{print}{fevardec}(
  x,
  digits = NULL,
  format = c("i|jz", "i|zj", "iz|j", "zi|j", "i|j|z"),
  ...
)

\method{print}{freqresp}(
  x,
  digits = NULL,
  format = c("i|jz", "i|zj", "iz|j", "zi|j", "i|j|z"),
  ...
)

\method{print}{spectrald}(
  x,
  digits = NULL,
  format = c("i|jz", "i|zj", "iz|j", "zi|j", "i|j|z"),
  ...
)
}
\arguments{
\item{x}{RLDM object, i.e. a \code{\link[=armamod]{armamod()}}, \code{\link[=rmfdmod]{rmfdmod()}}, \code{\link[=stspmod]{stspmod()}},
\code{\link[=impresp]{impresp()}}, \code{\link[=autocov]{autocov()}}, \code{\link[=freqresp]{freqresp()}},
\code{\link[=spectrum]{spectrum()}} or \code{\link[=fevardec]{fevardec()}} object.}

\item{digits}{(integer) if non \code{NULL} then correspondingly rounded numbers are printed,
see \code{\link[=round]{round()}}.}

\item{format}{(character string) selects specific output formats. Note that
\code{\link[=stsp]{stsp()}} and \code{\link[=fevardec]{fevardec()}} objects have no format option.
The option \code{'character'} is only implemented for (V)ARMA models.}

\item{...}{Further parameters are ignored.}
}
\value{
\code{invisible(x)}
}
\description{
Print Methods
}
\examples{
# for VARMA models six different print formats are implemented ###################
m = armamod(test_lmfd(dim = c(2,2), degrees = c(1,1)), sigma_L = diag(2))
print(m, digits = 2, format = "i|jz")
}
