% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLDM-package.R
\docType{package}
\name{RLDM-package}
\alias{RLDM}
\alias{RLDM-package}
\title{A Collection of Tools for VARMA and State Space Processes}
\description{
This package provides models for stationary processes with a rational spectral density and methods for their estimation.
}
\section{Author(s)}{


Wolfgang Scherrer and Bernd Funovits

Maintainer: \href{mailto:bernd.funovits@gmail.com}{bernd.funovits@gmail.com}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://bfunovits.github.io/RLDM}
  \item \url{https://github.com/bfunovits/RLDM}
  \item Report bugs at \url{https://github.com/bfunovits/RLDM/issues}
}

}
\author{
\strong{Maintainer}: Bernd Funovits \email{bernd.funovits@gmail.com} (\href{https://orcid.org/0000-0002-8247-6840}{ORCID})

Authors:
\itemize{
  \item Wolfgang Scherrer \email{wolfgang.scherrer@tuwien.ac.at}
}

}
\keyword{internal}
