% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLDM-package.R
\docType{package}
\name{RLDM}
\alias{RLDM}
\title{A Collection of Tools for VARMA and State Space Processes}
\description{
This package provides models for stationary processes with a rational spectral density and methods for their estimation.
}
\section{Author(s)}{


Wolfgang Scherrer and Bernd Funovits

Maintainer: \href{mailto:bernd.funovits@gmail.com}{bernd.funovits@gmail.com}
}

