% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_representations_classes.R
\name{rmfdmod}
\alias{rmfdmod}
\title{Constructor for RMFD Models}
\usage{
rmfdmod(sys, sigma_L = NULL, names = NULL, label = NULL)
}
\arguments{
\item{sys}{\code{\link[rationalmatrices:rmfd]{rationalmatrices::rmfd()}} object}

\item{sigma_L}{Left-factor of noise covariance,
i.e. the covariance \eqn{\sigma} is obtained as \code{sigma_L * t(sigma_L)}.
If \code{sigma_L} is a vector of dimension \eqn{n}, where \eqn{n} is the input dimension, only the diagonal elements are parametrized.
If it is a vector of dimension \eqn{n^2}, then the elements of \code{sigma_L} are filled column by column.}

\item{names}{optional vector of character strings}

\item{label}{optional character string}
}
\value{
Object of class \code{rmfdmod}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
A right-matrix fraction description (RMFD) plus parameterisation of noise covariance.
In \insertCite{Hannan.Deistler12}{RLDM}, RMFDs are also called dynamic adjustment forms.
Internally, MFDs are lists with slots \code{sys}, \code{sigma_L}, \code{names}, \code{label}.
\strong{Many of the generic functions which construct derived objects like the autocovariance \code{autocov()} are not yet implemented for \code{rmfdmod} objects.}
}
\examples{
y = rmfdmod(sys = test_rmfd(dim = c(3,2), degrees = c(2,2)))
y
}
\references{
\insertRef{Hannan.Deistler12}{RLDM}
}
