% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_utilities_data.R
\docType{data}
\name{RSdata}
\alias{RSdata}
\alias{RSdata_ts}
\alias{RSdata_xts}
\title{Ramey/Shapiro dataset}
\format{
A matrix with 248 rows and 7 variables (plus a timestamp, so in total 8 variables).
Versions as tibble (RSdata), ts (RSdata_ts), or xts (RSdata_xts) object are available:
\enumerate{
\item Date column
\item the logarithm of real per capita quantities of total government spending
\item real GDP
\item total hours worked
\item nondurable plus services consumption
\item private fixed investment
\item real wage (or more precisely the nominal compensation in private business divided by the deflator in private business)
\item tax rate
}

An object of class \code{mts} (inherits from \code{ts}, \code{matrix}) with 248 rows and 7 columns.

An object of class \code{xts} (inherits from \code{zoo}) with 248 rows and 7 columns.
}
\source{
\url{https://econweb.ucsd.edu/~vramey/research.html}
}
\usage{
RSdata

RSdata_ts

RSdata_xts
}
\description{
This dataset contains a (248 x 7)-dimensional matrix with quarterly US data from the first quarter
of 1947 till the last quarter of 2008.
The script for transforming the raw data (as csv-file) to the matrix RSdata is available
in the \emph{data-raw} directory.
The data set is used to identify government spending shocks and in particular to
investigate the response of consumption and real wages with respect to these shocks.
All variables are orthogonalized with respect to an intercept and a linear trend.
See \url{https://doi.org/10.1093/qje/qjq008} for more details.
}
\keyword{datasets}
