% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_utilities_str.R
\name{str methods}
\alias{str methods}
\alias{str.armamod}
\alias{str.stspmod}
\alias{str.impresp}
\alias{str.autocov}
\alias{str.fevardec}
\alias{str.freqresp}
\alias{str.spectrald}
\title{Display the Structure of Objects}
\usage{
\method{str}{armamod}(object, ...)

\method{str}{stspmod}(object, ...)

\method{str}{impresp}(object, ...)

\method{str}{autocov}(object, ...)

\method{str}{fevardec}(object, ...)

\method{str}{freqresp}(object, ...)

\method{str}{spectrald}(object, ...)
}
\arguments{
\item{object}{an object}

\item{...}{not used}
}
\value{
invisible(NULL)
}
\description{
Display the Structure of Objects
}
