% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_rldm_classes.R
\name{stspmod}
\alias{stspmod}
\title{Creator for stspmod class}
\usage{
stspmod(sys, sigma_L = NULL, names = NULL, label = NULL)
}
\arguments{
\item{sys}{\code{\link[rationalmatrices:stsp]{rationalmatrices::stsp()}} object}

\item{sigma_L}{noise covariance left}

\item{names}{optional vector of character strings}

\item{label}{optional chracter string}
}
\value{
Object of class \code{stspmod}.
}
\description{
Creator for stspmod class
}
\examples{
x = stspmod(sys = test_stsp(dim = c(2,2), s = 2), sigma_L = diag(2))
x
}
