% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afac_estimate_mom_ststp_subspace.R
\name{subspace order estimates}
\alias{subspace order estimates}
\alias{estorder_SVC}
\alias{estorder_IVC}
\alias{estorder_max}
\alias{estorder_rkH}
\alias{estorder_MOE}
\title{Helper Functions for Order Estimation}
\usage{
estorder_SVC(s.max, Hsv = NULL, n.par, n.obs, Hsize, penalty = "lnN", ...)

estorder_IVC(s.max, lndetSigma = NULL, n.par, n.obs, penalty = "BIC", ...)

estorder_max(s.max, ...)

estorder_rkH(s.max, Hsv = NULL, tol = sqrt(.Machine$double.eps), ...)

estorder_MOE(s.max, Hsv = NULL, ...)
}
\arguments{
\item{s.max}{(integer) maximum order (maximum state space dimension).}

\item{Hsv}{vector with the Hankel singular values (must have at least s.max entries).}

\item{n.par}{(s.max+1) dimensional vector with the respective number of parameters.}

\item{n.obs}{(integer) sample size \eqn{N} (or \code{Inf}).}

\item{Hsize}{two dimensional integer vector with the number of block rows and block columns
of the Hankel matrix (\code{Hsize = c(f,p)}).}

\item{penalty}{determines the penalty term. See the details below.}

\item{...}{optional additional parameters.}

\item{lndetSigma}{(s.max+1) dimensional vector with the logarithms of the determinants of the
respective estimated noise covariance matrices.}

\item{tol}{(small) tolarenace used to determine the rank of the Hankel matrix.}
}
\value{
Either NULL or a list with slots \verb{$s} (the selected/estimated order)
and \verb{$criterion} (an (\code{s.max+1}) dimensional vector).
}
\description{
These helper function are used in the subspace estimation routine
\link{subspace methods}
for the estimation of the order of a state space model.
For a discussion of order estimation in the context of
subspace methods see e.g \insertCite{Bauer2001}{RLDM}.
}
\details{
\code{estorder_max} simply returns the maximum order \code{s.max} considered.

\code{estorder_rkH} estimates the order by an estimate of the rank of the Hankel matrix.
If the maximum singular value is smaller than or equal to \code{.Machine$double.eps}
then the estimate is set to \code{s=0}. Otherwise the estimated order is equal to
the number of singular values which are larger than or equal to \code{tol} times the
maximum singular value.

The function \code{estorder_MOE}
searches for a "gap" in the singular values. The order is set to
maximum \eqn{s} which satisfies
\deqn{\ln(\sigma_s) > (\ln(\sigma_1)+\ln(\sigma_m))/2}{
      ln(\sigma[s]) > (ln(\sigma[1])+ln(\sigma[m]))/2}
where \eqn{\sigma_m}{\sigma[m]} is the minimum, non zero singular value.
This scheme is also implemented in the N4SID procedure of the
system identification toolbox of MATLAB (Ljung, 1991).

The function \code{estorder_SVC} implements the so called \emph{Singular Value Criterion}
\deqn{svc(s) = \sigma_{s+1}^2 + c(N)d(s)/N}{svc(s)=\sigma[s+1]^2 + c(N)d(s)/N}
(see e.g. \insertCite{Bauer2001}{RLDM}). Here \eqn{\sigma_s}{\sigma[s]} is the \eqn{s}-th
singular value of the weighted Hankel marix, \eqn{N} is the sample size,
\eqn{d(s) = 2ms} denotes the number of parameters for a state space
model with \eqn{s} states (and \eqn{m} outputs) and \eqn{c(N)} is a "penalty" term.
This term is
\eqn{c(N)=\ln(N)}{c(N)=ln(N)} for \code{penalty = "lnN"} and
\eqn{c(N)=fp\ln(N)}{c(N)=f p ln(N)} for \code{penalty = "fplnN"}. The estimate of the
order is the minimizer of this criterion.

\code{estorder_IVC} estimates the order via an information criterion of the form
\deqn{ivc(s) = \ln\det\hat\Sigma_{s} + c(N)d(s)/N}{ivc(s)=ln det \Sigma[s] + c(N)d(s)/N}
where \eqn{\hat\Sigma_s}{\Sigma[s]} is the estimate of the noise covariace matrix
obtained from a model with order \eqn{s}. Here the term \eqn{c(N)} is chosen as
\eqn{c(N)=2} for \code{penalty = "AIC"} and
\eqn{c(N)=\ln(N)}{c(N)=f p log(N)} for \code{penalty = "BIC"}.

Note also that for both routines \code{estorder_SVC} and \code{estorder_IVC}
one may also set \code{penalty} to an arbitrary numeric value! E.g.
setting \code{penalty=-1} would ensure that  \code{estorder_SVC}  always choses
the maximum possible order \code{s=s.max}.
}
\references{
\insertRef{Bauer2001}{RLDM}
}
