% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_compare_models.R
\name{compare_estimates}
\alias{compare_estimates}
\title{Compare Estimated Models}
\usage{
compare_estimates(estimates, y, n.lags = NULL)
}
\arguments{
\item{estimates}{(named) list of estimates. Each slot should contain a list with slots
\verb{$model} (the estimated model) and \verb{$n.par} the corresponding
number of (free) parameters of the model (class).}

\item{y}{(N-by-m)-dimensional matrix with the observed data (or an object which may
be coerced to a matrix with \code{as.matrix(y)}).}

\item{n.lags}{number of lags for the Portmantaeu test for serial correlation of the
residuals, see also \code{\link[=pm_test]{pm_test()}}.}
}
\value{
Matrix with the computed statistics for the estimated models. This matrix has attributes
\code{m}, \code{n.obs} and \code{n.lags}.
}
\description{
This utility function computes a number of statistics which may be used to compare/evaluate a set of
estimated models. In particular the function returns the log Likelihood (\code{ll}),
the Akaike Information Criterion (\code{AIC}), the Bayes Information Criterion (\code{BIC}),
the Final Prediction Error (\code{FPE}) and the p-values of a Portmanteau test for serial correlation
of the residuals.
}
\details{
The concentrated, conditional (scaled) log Likelihood
\deqn{ll = -(1/2)(m \ln(2\pi) + m + \ln\det S + 2 \ln\det (k_0)}{
      ll = -(1/2)(m ln(2\pi) + m + ln det S + 2 ln det (k[0])}
is computed with \code{ll(model, y, skip = 0, concentrated = TRUE)}, see \code{\link[=ll]{ll()}}.
Here \eqn{S} denotes the sample covariance of the residuals of the model.

The information criteria are
\deqn{AIC = -2 ll + (2/N) \kappa}
\deqn{BIC = -2 ll + (\ln(N)/N) \kappa}{BIC = -2 ll + (ln(N)/N) \kappa}
where \eqn{\kappa} denotes the respective number of free parameters.

The Final Prediction Error is
\deqn{FPE = \det(S)\frac{N+\kappa}{N-\kappa}}{FPE = det(S)(N+\kappa)/(N-\kappa)}

For the portmanteau test, see \code{\link[=pm_test]{pm_test()}}. If the number of lags is not
specified then the procedure choses a default value based on the sample size.
(This values is attached to the output as attribute).

Note that the procedure (re)evaulates these measures, even if the estimates contain
this information (e.g. the residuals or the log likelihood may be stored in the corresponding list).
The reason is to have a common data set and a common evaluation procedure for all estimates.

Typically the data \code{y} is the "estimation data set",
i.e. the data which has been used to estimate the models. However, one may also
pass a new data set to the procedure.
}
